# swagger.yaml
openapi: 3.0.1
info:
  title: ConfAPI for Confluence
  description: This resource provides methods for accessing user directory configuration
    for Confluence.
  contact:
    url: https://github.com/aservo/confluence-confapi-plugin
    email: info@aservo.com
  version: "1"
security:
- http_basic_scheme: []
paths:
...
  /directories:
    get:
      summary: Retrieves user directory information
      description: Upon successful request, returns a list of UserDirectoryBean object
        containing user directory details
      operationId: getDirectories
      responses:
        "200":
          description: user directory details list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserDirectoryBean'
        "400":
          description: An error occurred while retrieving the user directory list
    put:
      summary: Adds a new user directory
      description: Upon successful request, returns the added UserDirectoryBean object,
        Any existing configurations with the same name property are removed before
        adding the new configuration
      operationId: addDirectory
      parameters:
      - name: test
        in: query
        description: Whether or not to test the connection to the user directory service,
          e.g. CROWD (null defaults to TRUE)
        schema:
          type: boolean
      requestBody:
        description: The user directory to add
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserDirectoryBean'
        required: true
      responses:
        "200":
          description: user directory added
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserDirectoryBean'
        "400":
          description: An error occured while setting adding the new user directory
components:
  schemas:
...
    UserDirectoryBean:
      required:
      - appPassword
      - clientName
      - crowdUrl
      - implClass
      - name
      - type
      type: object
      properties:
        active:
          type: boolean
        name:
          maxLength: 2147483647
          minLength: 1
          type: string
        clientName:
          maxLength: 2147483647
          minLength: 1
          type: string
        type:
          type: string
          enum:
          - UNKNOWN
          - INTERNAL
          - CONNECTOR
          - CUSTOM
          - DELEGATING
          - CROWD
        description:
          type: string
        crowdUrl:
          maxLength: 2147483647
          minLength: 1
          type: string
        appPassword:
          maxLength: 2147483647
          minLength: 1
          type: string
        implClass:
          maxLength: 2147483647
          minLength: 1
          type: string
        proxyHost:
          type: string
        proxyPort:
          type: string
        proxyUsername:
          type: string
        proxyPassword:
          type: string
      xml:
        name: directory
  securitySchemes:
    http_basic_scheme:
      type: http